requirements.txt
flask>=3.0.0
flask-wtf>=1.2
flask-sqlalchemy>=3.1
wtforms-sqlalchemy>=0.4
markdown>=3.7
openai>=0.48.0
python-dotenv>=1.0.1